@@ -5,16 +5,12 @@ description = "Tool for classifying Boolean models by dynamic properties."
 authors = ["Ondřej Huvar <xhuvar@fi.muni.cz>", "Samuel Pastva <sam.pastva@gmail.com>"]
 license = "MIT"
 edition = "2021"
-rust-version = "1.60"
-
-[[bin]]
-name = "load-classifier-output"
-path = "src/bin/load_classifier_output.rs"
+rust-version = "1.72"
 
 [dependencies]
 biodivine-lib-bdd = ">=0.5.1, <1.0.0"
 biodivine-lib-param-bn = ">=0.4.5, <1.0.0"
-biodivine-hctl-model-checker = ">=0.1.3, <1.0.0"
+biodivine-hctl-model-checker = ">=0.1.7, <1.0.0"
 clap = { version = "4.1.1", features = ["derive"] }
 zip = "0.6.3"
 
