@@ -7,7 +7,7 @@ license = "MIT"
 repository = "https://github.com/sybila/biodivine-hctl-explorer"
 default-run = "hctl-property-explorer"
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.72"
 
 [build-dependencies]
 tauri-build = { version = "1.2.1", features = [] }
@@ -15,7 +15,7 @@ tauri-build = { version = "1.2.1", features = [] }
 [dependencies]
 biodivine-lib-bdd = ">=0.5.1, <1.0.0"
 biodivine-lib-param-bn = ">=0.4.5, <1.0.0"
-biodivine-hctl-model-checker = ">=0.1.3, <1.0.0"
+biodivine-hctl-model-checker = "0.1.7"
 
 rand = "0.8.5"
 clap = { version = "4.1.4", features = ["derive"] }
